project/wikimedia/requirements.txt
python-dotenv
confluent-kafka
opensearch-py
sseclient